Refactor playground web module by responsibility
